axi_pwm_gen: Cleanup
Because there are some major changes to the regmap
that break backwards compatibility, now is the time
to have an in-depth clean-up.
There are a few artefacts through the code which were
added in the initial architecture when the counter of
pwm 1 was used as a reference for other pwms when using
offsets. For this reason, we used intervals from 1 to 3,
and exposed each pwm counter for comparison with pwm 1 counter.

pause_cnt_d use was redundant

----------------------------------------------
The below section should be removed at squash.

We would like to avoid extra logic as much as possiblei.
Like in the regmap additions to the up_addr[3:0]+1
Which generates unnecessary logic.

Signed-off-by: AndreiGrozav <andrei.grozav@analog.com>